+++ test/local-payment/unit/internal/redirect-frame.js
@@ -32,7 +32,7 @@ describe('redirect-frame', function () {
 
     it('can put a redirect link onto the page if parent frame cannot be found and fallback is configured', function (done) {
       frameService.report.yieldsAsync(new Error('no frame'));
-      this.params.r = 'https://example.com/fallback-url';
+      this.params.r = global.encodeURIComponent('https://example.com/fallback-url');
       this.params.t = 'Return to Site';
 
       redirectFrame.start(function () {
@@ -60,7 +60,7 @@ describe('redirect-frame', function () {
 
     it('does not put a redirect link if text param is missing', function (done) {
       frameService.report.yieldsAsync(new Error('no frame'));
-      this.params.r = 'https://example.com/fallback-url';
+      this.params.r = global.encodeURIComponent('https://example.com/fallback-url');
 
       redirectFrame.start(function () {
         var link = document.querySelector('#container a');
@@ -73,7 +73,7 @@ describe('redirect-frame', function () {
 
     it('sanitizes fallback url', function (done) {
       frameService.report.yieldsAsync(new Error('no frame'));
-      this.params.r = 'javascript:alert("hey")'; // eslint-disable-line no-script-url
+      this.params.r = global.encodeURIComponent('javascript:alert("hey")'); // eslint-disable-line no-script-url
       this.params.t = 'Return to Site';
 
       redirectFrame.start(function () {